# Add VisitOperator trait and impls