… crates/wasmparser/src/binary_reader.rs
… crates/wasmparser/src/validator/func.rs
… crates/wasmparser/src/validator/operators.rs