Closes https://github.com/bytecodealliance/wasm-tools/issues/695.

@alexcrichton