@Robbepop 2022-08-03
Initial benchmarks indicate that somehow validation improved even though the current benchmarks do not even make use of the new visitor infrastructure. Instead we are still matching on `Operator` but it seems that even this part got slightly faster with this PR.
```
validate benchmark      time:   [11.015 ms 11.055 ms 11.105 ms]                               
                        change: [-7.9227% -7.0061% -6.1734%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe
```
Using the new visitor based validation we can see additional improvements:
```
validate benchmark      time:   [10.700 ms 10.745 ms 10.792 ms]                               
                        change: [-10.546% -9.6199% -8.7807%] (p = 0.00 < 0.05)
                        Performance has improved.
```