@Robbepop 2022-08-03
> This is looking good, thanks! One worry I have is that this is introducing two ways to do the same thing that we need to keep in sync an up-to-date. I think that the overhead here is somewhat low because it's easy to spot mistakes and such but I want to make sure that everything is at least thoroughly tested. For example I don't think there should be two different paths that read the binary encoding of an operator. The style implemented in the operator validator where `Operator` is deleted to visit methods I think makes sense.
> 
> I unfortunately can't open `validator/operators.rs` in a browser to provide more detailed feedback:
> 
>     * Could the `VisitOperator` trait have a method `visit_operator` which delegates to the appropriate `visit` method internally? This could then be used by the operator validator I believe.

This should totally work but kinda feels like it defeats the purpose of the `VisitOperator` trait. However, it should be fairly straight forward to provide it with a decent default implementation so that users can basically ignore it. Another potential solution is to get rid of the `Operator` `enum` entirely and do everything in terms of the new `VisitOperator` trait.

>     * Passing around `resources: &I` seems pretty onerous. If you'd like I think we could move ownership of the resources into the operator validator instead of having it separately owned in the function validator.

If we move ownership into the `OperatorValidator` then the only thing that distincts `OperatorValidator` from `FuncValidator` is an `index` field. Technically though it should work out when looking at the usage patterns of the `resources` in `FuncValidator`. Then we could make `VisitOperator<usize> impl for OperatorValidator` and put the entire error handling logic in there as well. I fear that the amount of refactoring work involved in this is going to be huge.

>     * If possible I'd ideally like to remove the dependence on `Operator` if we can. Given this PR is there any reason for it to stick around?

What do you mean by this? Remove the `Operator` `enum` entirely from the crate? If so: Yes I think we could potentially do that but then upgrading experience for current `wasmparser` users would be pretty bad. Maybe we could go a smooth transition using deprecation of the "old" techniques first and in the next release remove them? I personally don't mind a break on my end though since I would use the new APIs anyways.

> For benchmarks the in-repository benchmark isn't really that good I think, do you have more representative numbers on a more idiomatic module perhaps? I suspect that they're still wins, but it may be bigger than 10% since the modules in this repository are all test files which have few instructions.

In `wasmi` the biggest Wasm file we use for compilation and validation benchmarks is [this one here](https://github.com/robbepop/wasm_kernel/tree/6abd5b38d9775b3957087fd9b1e8cae9418602a0) compiled to Wasm.