@alexcrichton 2022-08-04
> This should totally work but kinda feels like it defeats the purpose of the VisitOperator trait

Ah this is just a convenience method if `Operator` sticks around, I don't expect high-perf cases to call it (e.g. validation). And yes this would be a default method so only one location would have to worry about defining it.

> Maybe we could go a smooth transition using deprecation of the "old" techniques first and in the next release remove them?

This crate does not provide a strong API stability guarantee, there's a whole lot of major versions and new things get added all the time. This basically isn't a factor when designing this crate. Instead we've been designing this crate as "keep it up to day with abstractions that all fit together", which is why given this trait I'm not sure we still need an `Operator` enum since it doesn't really make sense any more. If you'd prefer to leave it in that's ok because I agree it's work to remove, but I would at least want to try to remove duplication of decoding if not other duplicate pieces of functionality. Everything should still be tested in this crate at the very least.

> There benchmarks show that compiling and validating it using wasmi via wasmparser takes roughly 6ms which is a lot and also the reason why I started this investigation

Is that 6ms before this change or after this change? Are you able to see other bottlenecks in `wasmparser`?