@Robbepop 2022-08-04
> Ah this is just a convenience method if Operator sticks around, I don't expect high-perf cases to call it (e.g. validation). And yes this would be a default method so only one location would have to worry about defining it.

Will add the `visit_operator` method then. I don't plan on removing `Operator` just now. Instead I'd like to merge the PR and do that as a follow-up.

> Is that 6ms before this change or after this change? Are you able to see other bottlenecks in `wasmparser`?

In `wasmi` we unfortunately cannot use `wasmparser` crate direct but have to use a fork `wasmparser-nostd` for which I have not yet implemented this feature. Waiting for upstream to have it which is why I am behind this feature. ^^

I do see some bottlenecks with regards to small allocations but nothing I have benchmarked in detail, yet. Your concern about the double parsing is something where I have absolute no idea how we could improve the situation tbh.