@Robbepop 2022-08-04
@alexcrichton I moved the `WasmModuleResources` into the `OperatorValidator`. This cleaned up a lot of code in there and I am quite happy with the results. Unfortunately, the [`ModuleState::check_init_expr`](https://github.com/Robbepop/wasm-tools/blob/rf-visit-operator/crates/wasmparser/src/validator/core.rs#L321) method is unhappy about these changes and causes a borrow check error. I have not yet been able to fully resolve it since I am missing some information about the underlying behavior there. Do you have an idea how we should resolve this best?