@Robbepop 2022-08-07
I have implemented 2 different approaches to resolve the borrow checking issue mentioned [here](https://github.com/bytecodealliance/wasm-tools/pull/697#issuecomment-1205802898).

1. Implemented an `OperatorValidtorInner` which has no `resources` field and allow it to be temporarily transformed into a full `OperatorValidator` given some generic `resources: impl WasmModuleResources`. While this temporary conversion is costly it is only needed for validating const expressions. Therefore common validation of operators is still fast.
2. Implemented `OperatorValidator` with a `VisitOperator<(usize, T)>` pair where `T: WasmModuleResources`. While this makes the fowarding of parameters even worse than before this change it was possible to clean up some error handling parts with it and also it resolves the borrow checking issue.

The big issue with both approaches: Benchmarks show that both refactorings result in validation performance regressions from 10% down to 5% improvements which is significant.