@alexcrichton 2022-08-09
I checked out the commit prior to mine and I added a benchmark that additionally validates the entirety of [spidermonkey.wasm](https://github.com/bytecodealliance/sightglass/blob/main/benchmarks/spidermonkey/benchmark.wasm). Before after my PR to this PR I get:

| bench | before | after |
|------|-------|----------|
| validate | 11.9ms | 12.0ms |
| spidermonkey | 81.8ms | 79.2ms |

Taking a look at the profile I wrote this trivial patch:

```diff
diff --git a/crates/wasmparser/src/binary_reader.rs b/crates/wasmparser/src/binary_reader.rs
index f54a1414..983c0057 100644
--- a/crates/wasmparser/src/binary_reader.rs
+++ b/crates/wasmparser/src/binary_reader.rs
@@ -994,12 +994,19 @@ impl<'a> BinaryReader<'a> {
     ///
     /// If `BinaryReader` has no bytes remaining.
     pub fn read_u8(&mut self) -> Result<u8> {
-        self.ensure_has_byte()?;
-        let b = self.buffer[self.position];
+        let b = match self.buffer.get(self.position) {
+            Some(b) => *b,
+            None => return self.eof_err(),
+        };
         self.position += 1;
         Ok(b)
     }

+    #[cold]
+    fn eof_err<T>(&self) -> Result<T> {
+        Err(BinaryReaderError::eof(self.original_position(), 1))
+    }
+
     /// Advances the `BinaryReader` up to four bytes to parse a variable
     /// length integer as a `u32`.
     ///
```

and got this speedup:

| bench | before | after | with patch |
|------|-------|----------|-------|
| validate | 11.9ms | 12.0ms | 11.4ms |
| spidermonkey | 81.8ms | 79.2ms | 69.0ms |

Basically I can't really reproduce the numbers you're seeing. For performance work I think it would be best to settle on a set of benchmarks to validate (maybe `spidermonkey.wasm` is all that's needed to be added?) and go from there. Additionally I do not think that benchmarking should be done with LTO enabled. While I understand that it improves performance (and I measured it to improve perf locally) it doesn't reflect what consumers of the crate always use so it can paper over "easily missing" `#[inline]` annotations for example.

Otherwise I suspect there's quite a lot of low hanging fruit on `wasmparser` to improve performance, we've never really focused all that much on perf yet. Most of the perf of Wasmtime has been runtime execution since Cranelift takes awhile to compile things. We of course appreciate perf work on `wasmparser` though!

Are you able to quantify where the slowdown you're seeing is coming from? Did I miss something perhaps? I'm also on an aarch64 machine instead of an x86_64 machine which could make things different.